feature: cleanup of journal entry and activities
- Moved logic for adding activity entries to state into
ActivityEntrySubform because I think this makes more sense logically.
- Refactored UI component logic to be more declarative and reside near
the UI it controls for maintenance purposes.
- Built some types... still feeling out typescript but starting to like
it.

This should represent a pretty clear separation of logic for
controlling the various UI components more independently without prop
drilling or having to look through two components to see what one is
doing.